+++ src/main/java/io/github/ametsuchiru/enhancedannihilationplanes/EAPInterface.java
@@ -1,0 +1,9 @@
+package io.github.ametsuchiru.enhancedannihilationplanes;
+
+import net.minecraft.nbt.NBTTagCompound;
+
+public interface EAPInterface {
+
+    NBTTagCompound eap$writeEnchantments();
+
+}