+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Changelog
 
+## [2.0] - 2024-12-24
+
+### Added
+- Config options for enchantability, unbreaking chance, efficiency percentage, silk touch and fortune multipliers. All the options can be changed without restarting the game.
+- WAILA (HWYLA) Compatibility, planes in-world now show their enchantments in an aqua colour
+- Further optimizations, enchants are queried in a much more performant manner
+
+### Fixed
+- Wrenches now work
+
 ## [1.0] - 2023-09-21
 
 ### Changed